Streamline python package structure
The adsbibdesk.py at the root of the repo is now the adsbibdesk.py for
the python package itself. No more needing to 'build' the python
package. This was left-over from the days of injecting AppleScript into
Python code.

Now a user can just clone the repo and `python setup.py install` right
away!

closes #40