python setup.py service to run Service build
This setuptools command in setup.py runs the app build script, which
used to be in a separate build.py. Thus to build the automator action,
just run

python setup.py service

Also Add logging feedback to build.py: It's reassuring to know that
something is *actually* happening there.

As a result of doing this, it seems that the App is not getting updated;
only the service. I need to look at this.